% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inbox.R
\name{.count_attachments}
\alias{.count_attachments}
\title{Count Attachments}
\usage{
.count_attachments(EntryID, ol = NULL)
}
\arguments{
\item{EntryID}{entryID referring to an Outlook object}

\item{ol}{pointer to an active Outlook MAPI namespace}
}
\value{
integer
}
\description{
Count Attachments
}
