% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_array.R
\name{.format_table_array}
\alias{.format_table_array}
\title{Format Table Array}
\usage{
.format_table_array(tbl_ary)
}
\arguments{
\item{tbl_ary}{list object}
}
\value{
data.frame
}
\description{
Format Table Array
}
