% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_attachment.R
\name{.is_MailItem}
\alias{.is_MailItem}
\title{Is MailItem}
\usage{
.is_MailItem(x)
}
\arguments{
\item{x}{object(s) to be checked}
}
\value{
logical of length equal to the length of \code{x}
}
\description{
Returns \code{TRUE} if x is a pointer to a MailItem object, and \code{FALSE}
otherwise.
}
