% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inbox.R
\name{.lookup_exchange_sender}
\alias{.lookup_exchange_sender}
\title{Lookup Exchange Sender}
\usage{
.lookup_exchange_sender(EntryID, ol = NULL)
}
\arguments{
\item{EntryID}{entryID referring to an Outlook object}

\item{ol}{pointer to an active Outlook MAPI namespace}
}
\value{
character
}
\description{
This function is required because the email address of Exchange senders
aren't always displayed in the common \code{user@domail.com} form, and the SMTP
address form must be looked up for these users.
}
