% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{.make_outlook_address_list}
\alias{.make_outlook_address_list}
\title{Make Outlook Address List}
\usage{
.make_outlook_address_list(address)
}
\arguments{
\item{address}{character, vector of email addresses}
}
\value{
character singleton
}
\description{
Make Outlook Address List
}
