% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inbox.R
\name{.read_body}
\alias{.read_body}
\title{Read Body}
\usage{
.read_body(EntryID, ol = NULL)
}
\arguments{
\item{EntryID}{entryID referring to an Outlook object}

\item{ol}{pointer to an active Outlook MAPI namespace}
}
\value{
character
}
\description{
Read Body
}
