% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inbox.R
\name{read_inbox}
\alias{read_inbox}
\title{Read Inbox}
\usage{
read_inbox(folder_path, with_attachments_only = FALSE,
  count_attachments = FALSE, table_filter = NULL, whole_body = FALSE)
}
\arguments{
\item{folder_path}{character}

\item{with_attachments_only}{logical, should the results be filtered to include only emails with attachments?}

\item{count_attachments}{logical, should the number of attachments be returned? (Currently ignored)}

\item{table_filter}{character (optional), additional filters to be applied}

\item{whole_body}{logical, should the entire body of the email be returned? Default is FALSE (Currently ignored)}
}
\value{
data.frame
}
\description{
Read Inbox
}
