% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_attachment.R
\name{save_all_attachments}
\alias{save_all_attachments}
\title{Save All Attachments}
\usage{
save_all_attachments(mailItem, save_to_dir, must_work = FALSE)
}
\arguments{
\item{mailItem}{either a COMIDispatch pointer to a MailItem object, or an EntryID for a MailItem}

\item{save_to_dir}{character, path to the local directory where the files are to be saved}
}
\value{
character, vector of paths where attachments have been saved
}
\description{
Save All Attachments
}
