% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send Email}
\usage{
send_email(to, cc = NULL, subject, body, body_html = FALSE,
  attachments = NULL, display_only = FALSE)
}
\arguments{
\item{to}{character, vector of email addresses which will be receiving the email. Each element must contain the \code{@} symbol, with one or more characters on either side.}

\item{cc}{character (optional), vector of email addresses which will be carbon-copied (CCed) on the email. If provided, each element must contain the \code{@} symbol, with one or more characters on either side.}

\item{subject}{character, the subject line for the email. If this has length > 1, an error will be returned.}

\item{body}{character, the body of the email. If a vector of length > 1 is provided, the elements will be treated as lines of the email, and will be combined using \code{paste0(..., collapse = "\r\n")}.}

\item{body_html}{logical, does the argument to body contain HTML code that should control how the email is rendered? Default is \code{FALSE}.}

\item{attachments}{character (optional), paths to files that are to be attached. If any of these files does not exist, an error will be returned.}

\item{display_only}{logical, should the email be displayed only (and not sent)? Default is \code{FALSE}, which sends the email without displaying it first.}
}
\value{
NULL (invisibly)
}
\description{
Uses COM dispatch to communicate with MS Outlook and send an email.
}
