% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inbox.R
\name{.lookup_exchange_sender}
\alias{.lookup_exchange_sender}
\title{Lookup Exchange Sender}
\usage{
.lookup_exchange_sender(EntryID, ol = NULL)
}
\arguments{
\item{EntryID}{entryID referring to an Outlook object}

\item{ol}{pointer to an active Outlook MAPI namespace}
}
\value{
character
}
\description{
Lookup Exchange Sender
}
