% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_attachment.R
\name{save_attachment}
\alias{save_attachment}
\title{Save Attachment}
\usage{
save_attachment(mailItem, attachment, save_path, must_work = FALSE)
}
\arguments{
\item{mailItem}{either a COMIDispatch pointer to a MailItem object, or an EntryID for a MailItem}

\item{attachment}{the name (character) or index (integer) of the attachment to be saved}

\item{save_path}{character, path to the location where the file should be saved}
}
\value{
character, the path where the file was saved, or \code{NA_character_}
if a check determines that the file does not exist, meaning that the file was
not saved
}
\description{
Save a particular attachment to a location, given an identifier for the Email
the desired file is attached to, and some identifier for which attached file
to download.
}
